+++ src/main/java/com/wxl/dyttcrawler/web/service/DyttSearchService.java
@@ -121,7 +121,7 @@ public class DyttSearchService {
 
         }
 
-        return new Page<>(movies, (int) totalHits);
+        return new Page<>(movies, totalHits);
     }
 
     /**
